tommy 09/07/2021 12:41

+++ ft_print_xu.c
@@ -77,12 +77,12 @@ static void	ft_print_xu_alig (t_flags *flg, char c, LL num)
 
 void	ft_print_xu (t_flags *flg, va_list *pf_arg)
 {
-	char	c;
-	unsigned long int		num;
+	char			c;
+	unsigned int	num;
 
 	c = ' ';
 	flg->f_str = ft_strdup ("");
-	num = va_arg (*pf_arg, unsigned long int);
+	num = va_arg (*pf_arg, unsigned int);
 	if (num < 0)
 		flg->negative = true;
 	ft_print_xu_alig (flg, c, num);
